refactor: move parser and utils to separate folders
