+++ src/shared/api/clientApi/parsers/parseCategories.ts
