+++ src/shared/api/clientApi/parsers/parseProduct.ts
